requirements.txt
litellm
python-dotenv
requests
anthropic
openai